requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
aqueduct==1.11.2
asgiref==3.7.2
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
fastapi==0.110.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.0
h11==0.14.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
humanfriendly==10.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
image==1.5.33
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
nsfw-detector==1.1.1
NudeNet==2.0.9
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
open-nsfw-python3==0.0.5
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
priority==2.0.0
progressbar2==4.4.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydload==1.0.9
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.9
python-utils==3.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.7.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
